requirements.txt
gymnasium>=0.29
stable-baselines3>=2.3
torch>=2.2
scipy
numpy
pandas
matplotlib
seaborn
tensorboard
pyyaml
pytest
tqdm
